Refactoring and minor robustness improvements
