+++ Valheim.CustomRaids/Valheim.CustomRaids/Core/Configuration/ConfigurationLoader.cs
@@ -5,163 +5,162 @@ using System.IO;
 using System.Linq;
 using System.Text.RegularExpressions;
 
-namespace Valheim.CustomRaids.Core.Configuration
+namespace Valheim.CustomRaids.Core.Configuration;
+
+public static class ConfigurationLoader
 {
-    public static class ConfigurationLoader
+    private static readonly Regex SectionHeader = new Regex(@"(?<=[[]).+(?=[]])", RegexOptions.Compiled);
+    private static readonly Regex SectionSanitizer = new Regex(@"[^\p{L}\d.\[\]_\s]");
+    private static readonly Regex SectionWhiteSpaceTrim = new Regex(@"\s+(?=\])");
+
+    /// <summary>
+    /// Attempts to clean up section headers in file.
+    /// </summary>
+    public static void SanitizeSectionHeaders(string filePath)
     {
-        private static readonly Regex SectionHeader = new Regex(@"(?<=[[]).+(?=[]])", RegexOptions.Compiled);
-        private static readonly Regex SectionSanitizer = new Regex(@"[^\p{L}\d.\[\]_\s]");
-        private static readonly Regex SectionWhiteSpaceTrim = new Regex(@"\s+(?=\])");
-
-        /// <summary>
-        /// Attempts to clean up section headers in file.
-        /// </summary>
-        public static void SanitizeSectionHeaders(string filePath)
+        if (File.Exists(filePath))
         {
-            if (File.Exists(filePath))
+            var lines = File.ReadAllLines(filePath);
+
+            for (int i = 0; i < lines.Length; ++i)
             {
-                var lines = File.ReadAllLines(filePath);
+                string line = lines[i];
 
-                for (int i = 0; i < lines.Length; ++i)
+                if (SectionHeader.IsMatch(line))
                 {
-                    string line = lines[i];
-
-                    if (SectionHeader.IsMatch(line))
-                    {
-                        var sanitized = SectionSanitizer.Replace(line, "");
-                        sanitized = SectionWhiteSpaceTrim.Replace(sanitized, "");
-                        lines[i] = sanitized;
-                    }
+                    var sanitized = SectionSanitizer.Replace(line, "");
+                    sanitized = SectionWhiteSpaceTrim.Replace(sanitized, "");
+                    lines[i] = sanitized;
                 }
-
-                File.WriteAllLines(filePath, lines);
             }
+
+            File.WriteAllLines(filePath, lines);
+        }
+    }
+
+    public static TConfig LoadConfiguration<TConfig>(ConfigFile configFile) where TConfig : Config, IConfigFile
+    {
+        var mainConfig = Activator.CreateInstance<TConfig>();
+
+        if (string.IsNullOrEmpty(configFile.ConfigFilePath))
+        {
+            Log.LogError("Unable to load unloaded ConfigFile.");
+            return mainConfig;
         }
 
-        public static TConfig LoadConfiguration<TConfig>(ConfigFile configFile) where TConfig : Config, IConfigFile
+        //Scan for headers
+        var sectionHeaders = ScanSectionHeads(configFile.ConfigFilePath);
+
+        foreach (var sectionHead in sectionHeaders)
         {
-            var mainConfig = Activator.CreateInstance<TConfig>();
+            Log.LogTrace($"Loading config entries for '{sectionHead}'.");
 
-            if (string.IsNullOrEmpty(configFile.ConfigFilePath))
+            //Identify header components
+            var components = sectionHead.Split(new[] { '.' }, StringSplitOptions.RemoveEmptyEntries);
+
+            if (components.Length < 1 || string.IsNullOrEmpty(components[0]))
             {
-                Log.LogError("Unable to load unloaded ConfigFile.");
-                return mainConfig;
-            }
-
-            //Scan for headers
-            var sectionHeaders = ScanSectionHeads(configFile.ConfigFilePath);
-
-            foreach (var sectionHead in sectionHeaders)
-            {
-                Log.LogTrace($"Loading config entries for '{sectionHead}'.");
-
-                //Identify header components
-                var components = sectionHead.Split(new[] { '.' }, StringSplitOptions.RemoveEmptyEntries);
-
-                if (components.Length < 1 || string.IsNullOrEmpty(components[0]))
-                {
 #if DEBUG
                     Log.LogTrace("Early stop. 0 components found.");
 #endif
 
-                    continue;
-                }
-
-                LoadRecursively(configFile, mainConfig, components, components[0], 0);
+                continue;
             }
 
-            return mainConfig;
+            LoadRecursively(configFile, mainConfig, components, components[0], 0);
         }
 
-        private static void LoadRecursively<T>(ConfigFile configFile, T currentConfig, IList<string> sectionParts, string sectionKey, int depth) where T : Config
-        {
-            currentConfig.SectionName = sectionParts[depth];
-            currentConfig.SectionKey = sectionKey;
+        return mainConfig;
+    }
 
-            //Check if we have reached the end
-            if (depth == sectionParts.Count - 1 && currentConfig is not IConfigFile)
-            {
+    private static void LoadRecursively<T>(ConfigFile configFile, T currentConfig, IList<string> sectionParts, string sectionKey, int depth) where T : Config
+    {
+        currentConfig.SectionName = sectionParts[depth];
+        currentConfig.SectionKey = sectionKey;
+
+        //Check if we have reached the end
+        if (depth == sectionParts.Count - 1 && currentConfig is not IConfigFile)
+        {
 #if DEBUG
                 Log.LogTrace($"Binding entries for {sectionKey}:{currentConfig.GetType().Name}");
 #endif
-                BindObjectEntries(configFile, currentConfig, sectionKey);
-                return;
-            }
-
-            //Check if we are at the top level. If so, don't increment subsections.
-            if (currentConfig is IConfigFile topConfig)
-            {
-                var subsection = topConfig.GetSubsection(sectionKey);
-
-                LoadRecursively(configFile, subsection, sectionParts, sectionKey, depth);
-            }
-            //Go deeper into tree.
-            else if (currentConfig is IHaveSubsections parent)
-            {
-                var subsection = parent.GetSubsection(sectionParts[depth + 1]);
-                var subsectionKey = sectionKey + "." + sectionParts[depth + 1];
-
-                LoadRecursively(configFile, subsection, sectionParts, subsectionKey, depth + 1);
-            }
+            BindObjectEntries(configFile, currentConfig, sectionKey);
+            return;
         }
 
-        private static void BindObjectEntries<T>(ConfigFile configFile, T config, string sectionHeader) where T : Config
+        //Check if we are at the top level. If so, don't increment subsections.
+        if (currentConfig is IConfigFile topConfig)
         {
-            Type entryType = typeof(IConfigurationEntry);
+            var subsection = topConfig.GetSubsection(sectionKey);
 
-            var fields = config
-                    .GetType()
-                    .GetFields()
-                    .Where(x => entryType.IsAssignableFrom(x.FieldType))
-                    .ToList();
-
-            foreach (var field in fields)
-            {
-                Log.LogTrace($"Creating and binding entry for '{sectionHeader}:{field.Name}'");
-
-                var entry = field.GetValue(config) as IConfigurationEntry;
-
-                if (entry is null)
-                {
-                    entry = (IConfigurationEntry)Activator.CreateInstance(field.FieldType);
-                    field.SetValue(config, entry);
-                }
-
-                entry.Bind(configFile, sectionHeader, field.Name);
-
-                Log.LogTrace($"[{sectionHeader}]: Loaded {field.Name}:{entry}");
-            }
+            LoadRecursively(configFile, subsection, sectionParts, sectionKey, depth);
         }
-
-        private static List<string> ScanSectionHeads(string configFile)
+        //Go deeper into tree.
+        else if (currentConfig is IHaveSubsections parent)
         {
-            if (!File.Exists(configFile))
-            {
-                return new List<string>();
-            }
+            var subsection = parent.GetSubsection(sectionParts[depth + 1]);
+            var subsectionKey = sectionKey + "." + sectionParts[depth + 1];
 
-            Log.LogTrace($"Scanning config sections in {configFile}");
-
-            //Scan for headers
-            var lines = File.ReadAllLines(configFile);
-
-            var sectionHeaders = new List<string>(lines.Count() / 5); //Just some random guess at lines pr header.
-
-            foreach (var line in lines)
-            {
-                var sectionMatch = SectionHeader.Match(line);
-
-                if (sectionMatch.Success)
-                {
-                    var sectionHeader = sectionMatch.Value;
-
-                    Log.LogTrace($"Found section '{sectionHeader}'");
-
-                    sectionHeaders.Add(sectionHeader);
-                }
-            }
-
-            return sectionHeaders;
+            LoadRecursively(configFile, subsection, sectionParts, subsectionKey, depth + 1);
         }
+    }
+
+    private static void BindObjectEntries<T>(ConfigFile configFile, T config, string sectionHeader) where T : Config
+    {
+        Type entryType = typeof(IConfigurationEntry);
+
+        var fields = config
+                .GetType()
+                .GetFields()
+                .Where(x => entryType.IsAssignableFrom(x.FieldType))
+                .ToList();
+
+        foreach (var field in fields)
+        {
+            Log.LogTrace($"Creating and binding entry for '{sectionHeader}:{field.Name}'");
+
+            var entry = field.GetValue(config) as IConfigurationEntry;
+
+            if (entry is null)
+            {
+                entry = (IConfigurationEntry)Activator.CreateInstance(field.FieldType);
+                field.SetValue(config, entry);
+            }
+
+            entry.Bind(configFile, sectionHeader, field.Name);
+
+            Log.LogTrace($"[{sectionHeader}]: Loaded {field.Name}:{entry}");
+        }
+    }
+
+    private static List<string> ScanSectionHeads(string configFile)
+    {
+        if (!File.Exists(configFile))
+        {
+            return new List<string>();
+        }
+
+        Log.LogTrace($"Scanning config sections in {configFile}");
+
+        //Scan for headers
+        var lines = File.ReadAllLines(configFile);
+
+        var sectionHeaders = new List<string>(lines.Count() / 5); //Just some random guess at lines pr header.
+
+        foreach (var line in lines)
+        {
+            var sectionMatch = SectionHeader.Match(line);
+
+            if (sectionMatch.Success)
+            {
+                var sectionHeader = sectionMatch.Value;
+
+                Log.LogTrace($"Found section '{sectionHeader}'");
+
+                sectionHeaders.Add(sectionHeader);
+            }
+        }
+
+        return sectionHeaders;
     }
 }
