+++ Valheim.CustomRaids/Valheim.CustomRaids/Patches/RandEventSystemHaveGlobalKeyPatch.cs
@@ -1,49 +1,46 @@
 ﻿using HarmonyLib;
 using System.Collections.Generic;
-using UnityEngine;
 using Valheim.CustomRaids.Compatibilities;
 using Valheim.CustomRaids.Conditions;
-using Valheim.CustomRaids.Core;
 
-namespace Valheim.CustomRaids.Patches
+namespace Valheim.CustomRaids.Patches;
+
+[HarmonyPatch(typeof(RandEventSystem))]
+internal static class RandEventSystemHaveGlobalKeyPatch
 {
-    [HarmonyPatch(typeof(RandEventSystem))]
-    internal static class RandEventSystemHaveGlobalKeyPatch
+    [HarmonyPatch(nameof(RandEventSystem.HaveGlobalKeys))]
+    [HarmonyPostfix]
+    private static void IgnoreIfEnhancedKeysIsInstalled(RandomEvent ev, ref bool __result)
     {
-        [HarmonyPatch("HaveGlobalKeys")]
-        [HarmonyPostfix]
-        private static void IgnoreIfEnhancedKeysIsInstalled(RandomEvent ev, ref bool __result)
-        {
-            __result = true;
-        }
+        __result = true;
+    }
 
-        [HarmonyPatch("GetValidEventPoints")]
-        [HarmonyPrefix]
-        private static void FilterCharactersWithInvalidGlobalKeys(RandomEvent ev, ref List<ZDO> characters)
-        {
-            List<ZDO> filtered = new List<ZDO>();
+    [HarmonyPatch(nameof(RandEventSystem.GetValidEventPoints))]
+    [HarmonyPrefix]
+    private static void FilterCharactersWithInvalidGlobalKeys(RandomEvent ev, ref List<ZDO> characters)
+    {
+        List<ZDO> filtered = new List<ZDO>();
 
-            for(int i = 0; i < characters.Count; ++i)
+        for (int i = 0; i < characters.Count; ++i)
+        {
+
+            var playerName = EnhancedProgressTrackerCompatibilities.GetPlayerName(characters[i]);
+
+            if (playerName is not null)
             {
-
-                var playerName = EnhancedProgressTrackerCompatibilities.GetPlayerName(characters[i]);
-
-                if (playerName is not null)
+                if (GlobalKeyConditionChecker.ShouldFilter(ev, playerName))
                 {
-                    if(GlobalKeyConditionChecker.ShouldFilter(ev, playerName))
-                    {
 #if DEBUG
                         Log.LogDebug($"Filtering GetValidEventPoints for {ev.m_name}");
 #endif
 
-                        continue;
-                    }
+                    continue;
                 }
-
-                filtered.Add(characters[i]);
             }
 
-            characters = filtered;
+            filtered.Add(characters[i]);
         }
+
+        characters = filtered;
     }
 }
