+++ Valheim.CustomRaids/Valheim.CustomRaids/Patches/RandEventSystemPatch.cs
@@ -1,29 +1,27 @@
 ﻿using HarmonyLib;
 using Valheim.CustomRaids.Core;
-using Valheim.CustomRaids.Patches;
 using Valheim.CustomRaids.Raids.Managers;
 
-namespace Valheim.CustomRaids
-{
-    [HarmonyPatch(typeof(RandEventSystem), "Start")]
-    internal static class RandEventSystemPatch
-    {
-        [HarmonyPostfix]
-        private static void RandEventSystemStart(RandEventSystem __instance)
-        {
-            Log.LogDebug("Starting RandEventSystem");
+namespace Valheim.CustomRaids.Patches;
 
-            //If singleplayer, ZNet will not be initialized here.
-            if (ZNet.instance == null)
-            {
-                RaidConfigManager.ApplyConfigs();
-                RandEventSystemWaitPatch.Wait = false;
-            }
-            else if(ZNet.instance.IsServer())
-            {
-                RaidConfigManager.ApplyConfigs();
-                RandEventSystemWaitPatch.Wait = false;
-            }
+[HarmonyPatch(typeof(RandEventSystem), nameof(RandEventSystem.Start))]
+internal static class RandEventSystemPatch
+{
+    [HarmonyPostfix]
+    private static void RandEventSystemStart(RandEventSystem __instance)
+    {
+        Log.LogDebug("Starting RandEventSystem");
+
+        //If singleplayer, ZNet will not be initialized here.
+        if (ZNet.instance == null)
+        {
+            RaidConfigManager.ApplyConfigs();
+            RandEventSystemWaitPatch.Wait = false;
+        }
+        else if (ZNet.instance.IsServer())
+        {
+            RaidConfigManager.ApplyConfigs();
+            RandEventSystemWaitPatch.Wait = false;
         }
     }
 }
