+++ Valheim.CustomRaids/Valheim.CustomRaids/Patches/SpawnPatch.cs
@@ -1,26 +1,25 @@
 ﻿using HarmonyLib;
 
-namespace Valheim.CustomRaids.Patches
+namespace Valheim.CustomRaids.Patches;
+
+[HarmonyPatch(typeof(Character))]
+internal static class CharacterSetFactionPatch
 {
-    [HarmonyPatch(typeof(Character))]
-    internal static class CharacterSetFactionPatch
+    [HarmonyPatch(nameof(Character.Start))]
+    [HarmonyPostfix]
+    private static void AssignFaction(Character __instance, ZNetView ___m_nview)
     {
-        [HarmonyPatch("Start")]
-        [HarmonyPostfix]
-        private static void AssignFaction(Character __instance, ZNetView ___m_nview)
+        var zdo = ___m_nview?.GetZDO();
+        if (zdo is null)
         {
-            var zdo = ___m_nview?.GetZDO();
-            if (zdo is null)
-            {
-                return;
-            }
+            return;
+        }
 
-            var faction = zdo.GetInt("faction", -1);
+        var faction = zdo.GetInt("faction", -1);
 
-            if (faction >= 0)
-            {
-                __instance.m_faction = (Character.Faction)faction;
-            }
+        if (faction >= 0)
+        {
+            __instance.m_faction = (Character.Faction)faction;
         }
     }
 }
